Started Material UI Notes
Started taking notes on Material UI library since it's a leading component library. Note that this is low priority as we've learned bootstrap and tailwind, but it's still something to take into consideration. Now we're probably going to mainly focus on moving forward with learning back-end with the odin project.