Finished MongoDB notes
Finished the MongoDB notes so I have a pretty good grasp of it now. Now it's time to move on with the course and learn about Express. We've already been exposed to a little bit of its syntax and logic, but we'll create a whole separate notes section for express